+++ base/src/test/kotlin/proguard/analysis/cpa/CfaUtilTest.kt
@@ -33,7 +33,8 @@ class CfaUtilTest : FreeSpec({
                         }
                     }
                     """.trimIndent()
-                )
+                ),
+                javacArguments = listOf("-source", "1.8", "-target", "1.8")
             ).programClassPool
         ) { false }