+++ base/src/test/kotlin/proguard/analysis/cpa/JvmTaintCpaTest.kt
@@ -99,7 +99,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -144,7 +145,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -200,7 +202,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -254,7 +257,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -296,7 +300,8 @@ class JvmTaintCpaTest : FreeSpec({
                     }
                 }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -345,7 +350,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -394,7 +400,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -450,7 +457,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -511,7 +519,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -569,7 +578,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -607,7 +617,8 @@ class JvmTaintCpaTest : FreeSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature