+++ base/src/test/kotlin/proguard/analysis/cpa/TraceExtractorTest.kt
@@ -118,7 +118,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -171,7 +172,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -225,7 +227,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -297,7 +300,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -349,7 +353,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -396,7 +401,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -448,7 +454,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -496,7 +503,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -543,7 +551,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -590,7 +599,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -645,7 +655,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -720,7 +731,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -772,7 +784,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -822,7 +835,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -915,7 +929,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -995,7 +1010,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -1080,7 +1096,8 @@ class TraceExtractorTest : StringSpec({
                         }
                     }
                         """.trimIndent()
-                    )
+                    ),
+                    javacArguments = listOf("-source", "1.8", "-target", "1.8")
                 ).programClassPool
             )
             val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature