+++ base/src/test/kotlin/proguard/analysis/cpa/TreeHeapTest.kt
@@ -127,7 +127,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -173,7 +174,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -219,7 +221,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -267,7 +270,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -317,7 +321,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -363,7 +368,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -412,7 +418,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -462,7 +469,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -507,7 +515,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -557,7 +566,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -636,7 +646,8 @@ class TreeHeapTest : StringSpec({
                         }
                     }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature
@@ -714,7 +725,8 @@ class TreeHeapTest : StringSpec({
                     }                
                 }
                                     """.trimIndent()
-                                )
+                                ),
+                                javacArguments = listOf("-source", "1.8", "-target", "1.8")
                             ).programClassPool
                         )
                         val mainSignature = interproceduralCfa!!.functionEntryNodes.stream().filter { it.signature.fqn.contains("main") }.findFirst().get().signature