+++ base/src/test/kotlin/proguard/classfile/editor/AccessFixerTest.kt
@@ -28,7 +28,8 @@ class AccessFixerTest : FreeSpec({
                         public final void foo() { }
                     }
                 """.trimIndent()
-            )
+            ),
+            javacArguments = listOf("-source", "1.8", "-target", "1.8")
         )
 
         "Then the super method access flags should not be modified incorrectly" {
@@ -68,7 +69,8 @@ class AccessFixerTest : FreeSpec({
                         public final void foo() { }
                     }
                 """.trimIndent()
-            )
+            ),
+            javacArguments = listOf("-source", "1.8", "-target", "1.8")
         )
 
         "When the referencing class is repackaged" - {